missing_shirli_scripts/requirements.txt
gspread>=5.7.0
google-auth>=2.16.0
google-auth-oauthlib>=0.8.0
google-auth-httplib2>=0.1.0
spotipy>=2.22.1
google-api-python-client>=2.70.0
